Switching to supertab autocomplete plugin
The reason is that the order of the complete items from the c-n
autocomplete method just never made sense to me. It seems that there are
some limitations in terms of vim because it performs a linear scan
through the files. Still it seems that when I type a partial match, my
target is always at the bottom in my projects while the longest match is
at the top, which turned out to be unbearable after some time.

Also looked at YouCompleteMe but seems way too complex.
Supertab seems to be the next simplest choice by the community.